+++ src/controlledEntityTreeWidget.cpp
@@ -28,18 +28,21 @@
 #include <QHeaderView>
 #include <QMenu>
 
+#include <unordered_set>
+
+enum class Kind
+{
+	None,
+	EntityModelNode,
+	VirtualNode,
+};
+
 // Base node item
 class NodeItem : public QObject, public QTreeWidgetItem
 {
 	using QObject::parent;
 
 public:
-	enum class Kind
-	{
-		EntityModelNode,
-		VirtualNode,
-	};
-
 	Kind kind() const
 	{
 		return _kind;
@@ -63,7 +66,7 @@ public:
 		// Also update the parent node
 		if (auto* parent = static_cast<NodeItem*>(QTreeWidgetItem::parent()))
 		{
-			if (parent->kind() == NodeItem::Kind::VirtualNode)
+			if (parent->kind() == Kind::VirtualNode)
 			{
 				parent->updateHasError();
 			}
@@ -87,9 +90,9 @@ protected:
 	virtual void updateHasError() {}
 
 private:
-	la::avdecc::entity::model::DescriptorType const _descriptorType;
-	la::avdecc::entity::model::DescriptorIndex const _descriptorIndex;
-	Kind const _kind;
+	la::avdecc::entity::model::DescriptorType const _descriptorType{ la::avdecc::entity::model::DescriptorType::Invalid };
+	la::avdecc::entity::model::DescriptorIndex const _descriptorIndex{ 0u };
+	Kind const _kind{ Kind::None };
 	bool _hasError{ false };
 };
 
@@ -153,6 +156,30 @@ protected:
 class ControlledEntityTreeWidgetPrivate : public QObject, public la::avdecc::controller::model::EntityModelVisitor
 {
 public:
+	struct NodeIdentifier
+	{
+		la::avdecc::entity::model::DescriptorType type{ la::avdecc::entity::model::DescriptorType::Invalid };
+		la::avdecc::entity::model::DescriptorIndex index{ 0u };
+		Kind kind{ Kind::None };
+
+		constexpr bool operator==(NodeIdentifier const& other) const noexcept
+		{
+			return type == other.type && index == other.index && kind == other.kind;
+		}
+
+		/** Hash functor to be used for std::hash */
+		struct hash
+		{
+			std::size_t operator()(NodeIdentifier const& id) const
+			{
+				// We use 16 bits for the descriptor type, 15 bits for the index and 1 for the kind
+				return (static_cast<std::size_t>(id.type) << 16) | (static_cast<std::size_t>(id.index & 0x7fff) << 1) | static_cast<std::size_t>(id.kind);
+			}
+		};
+	};
+
+	using NodeIdentifierSet = std::unordered_set<NodeIdentifier, NodeIdentifier::hash>;
+
 	ControlledEntityTreeWidgetPrivate(ControlledEntityTreeWidget* q)
 		: q_ptr(q)
 	{
@@ -170,8 +197,6 @@ public:
 
 		q->setControlledEntityID(la::avdecc::UniqueIdentifier{});
 		q->clearSelection();
-
-		_entityExpandedStates.clear();
 	}
 
 	Q_SLOT void entityOnline(la::avdecc::UniqueIdentifier const entityID)
@@ -186,16 +211,12 @@ public:
 
 	Q_SLOT void entityOffline(la::avdecc::UniqueIdentifier const entityID)
 	{
-		// Always remove saved state
-		_entityExpandedStates.erase(entityID);
-
-		if (_controlledEntityID != entityID)
+		// The current entity went offline, clear everything
+		if (_controlledEntityID == entityID)
 		{
-			return;
+			Q_Q(ControlledEntityTreeWidget);
+			q->clearSelection();
 		}
-
-		Q_Q(ControlledEntityTreeWidget);
-		q->clearSelection();
 	}
 
 	Q_SLOT void streamInputErrorCounterChanged(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::DescriptorIndex const descriptorIndex, la::avdecc::entity::StreamInputCounterValidFlags const flags)
@@ -205,64 +226,76 @@ public:
 			return;
 		}
 
-		if (auto* item = findEntityModelNodeItem(la::avdecc::entity::model::DescriptorType::StreamInput, descriptorIndex))
+		if (auto* item = findItem({ la::avdecc::entity::model::DescriptorType::StreamInput, descriptorIndex }))
 		{
 			item->setHasError(!flags.empty());
 		}
 	}
 
-	void saveSelectedDescriptor()
-	{
-		// TODO
-	}
-
-	void restoreSelectedDescriptor()
-	{
-		Q_Q(ControlledEntityTreeWidget);
-
-		// TODO: Properly restore the previous saved index - Right now always restore the first one
-		q->setCurrentIndex(q->model()->index(0, 0));
-	}
-
-	void saveExpandedState()
+	void saveUserTreeWidgetState()
 	{
 		// Build expanded state
-		NodeExpandedStates states;
+		auto currentNode = NodeIdentifier{};
+		auto expandedNodes = NodeIdentifierSet{};
 
 		Q_Q(ControlledEntityTreeWidget);
-		for (auto const& kv : _map)
+		for (auto const& [id, item] : _identifierToNodeItem)
 		{
-			auto const& node{ kv.first };
-			auto const& item{ kv.second };
-			states.insert({ node, q->isItemExpanded(item) });
+			if (item == q->currentItem())
+			{
+				currentNode = id;
+			}
+
+			// Put only the expanded nodes in the set
+			if (q->isItemExpanded(item))
+			{
+				expandedNodes.insert(id);
+			}
+		}
+
+		if (_controlledEntityID.getValue() == 0x001B92FFFE01B930)
+		{
+			_userTreeWidgetStates[la::avdecc::UniqueIdentifier{ 0x001B92FFFF01B930 }] = { currentNode, expandedNodes };
+		}
+		if (_controlledEntityID.getValue() == 0x001B92FFFF01B930)
+		{
+			_userTreeWidgetStates[la::avdecc::UniqueIdentifier{ 0x001B92FFFE01B930 }] = { currentNode, expandedNodes };
 		}
 
 		// Save expanded state for previous EntityID
-		_entityExpandedStates[_controlledEntityID] = states;
+		_userTreeWidgetStates[_controlledEntityID] = { currentNode, std::move(expandedNodes) };
 	}
 
-	void restoreExpandedState()
+	void restoreUserTreeWidgetState()
 	{
 		Q_Q(ControlledEntityTreeWidget);
 
-		// Load expanded state if any
-		auto const statesIt = _entityExpandedStates.find(_controlledEntityID);
-		if (statesIt != _entityExpandedStates.end())
+		auto nodeSelected = false;
+		auto const it = _userTreeWidgetStates.find(_controlledEntityID);
+		if (it != std::end(_userTreeWidgetStates))
 		{
-			// Restore expanded state
-			auto const& states = statesIt->second;
-			for (auto const& kv : _map)
+			auto const& userTreeWidgetState = it->second;
+			for (auto const& id : userTreeWidgetState.expandedNodes)
 			{
-				auto const& node{ kv.first };
-				auto const& item{ kv.second };
-				try
+				if (auto* nodeItem = findItem(id))
 				{
-					q->setItemExpanded(item, states.at(node));
-				}
-				catch (...)
-				{
+					nodeItem->setExpanded(true);
 				}
 			}
+
+			if (auto* selectedItem = findItem(userTreeWidgetState.currentNode))
+			{
+				auto const index = q->indexFromItem(selectedItem);
+				q->setCurrentIndex(index);
+				nodeSelected = true;
+			}
+		}
+
+		// First time we see this entity or model changed
+		if (!nodeSelected)
+		{
+			// Select the first node, which is always Entity Descriptor
+			q->setCurrentIndex(q->model()->index(0, 0));
 		}
 	}
 
@@ -271,10 +304,12 @@ public:
 		Q_Q(ControlledEntityTreeWidget);
 
 		q->clear();
-		_map.clear();
+		_identifierToNodeItem.clear();
 
 		if (!_controlledEntityID)
+		{
 			return;
+		}
 
 		auto& manager = avdecc::ControllerManager::getInstance();
 		auto controlledEntity = manager.getControlledEntity(_controlledEntityID);
@@ -285,10 +320,7 @@ public:
 		}
 
 		// Restore expanded state for new EntityID
-		restoreExpandedState();
-
-		// Restore selected descriptor for new EntityID
-		restoreSelectedDescriptor();
+		restoreUserTreeWidgetState();
 	}
 
 	void setControlledEntityID(la::avdecc::UniqueIdentifier const entityID)
@@ -302,8 +334,7 @@ public:
 
 		if (_controlledEntityID)
 		{
-			saveExpandedState();
-			saveSelectedDescriptor();
+			saveUserTreeWidgetState();
 		}
 
 		_controlledEntityID = entityID;
@@ -316,48 +347,38 @@ public:
 		return _controlledEntityID;
 	}
 
-	NodeItem* findEntityModelNodeItem(la::avdecc::entity::model::DescriptorType const& descriptorType, la::avdecc::entity::model::DescriptorIndex const descriptorIndex) const
+	NodeItem* findItem(NodeIdentifier const& nodeIdentifier) const
 	{
-		auto const it = std::find_if(std::begin(_map), std::end(_map),
-			[&descriptorType, &descriptorIndex](auto const& kv)
+		auto const it = _identifierToNodeItem.find(nodeIdentifier);
+		return it != std::end(_identifierToNodeItem) ? it->second : nullptr;
+	}
+
+	NodeIdentifier findNodeIdentifier(NodeItem const* item) const
+	{
+		auto const it = std::find_if(std::begin(_identifierToNodeItem), std::end(_identifierToNodeItem),
+			[item](auto const& kv)
 			{
-				auto const* item = kv.second;
-				if (item->kind() == NodeItem::Kind::VirtualNode)
-				{
-					// Exclude virtual nodes
-					return false;
-				}
-				else
-				{
-					return item->descriptorType() == descriptorType && item->descriptorIndex() == descriptorIndex;
-				}
+				return kv.second == item;
 			});
-		if (it != std::end(_map))
-		{
-			return it->second;
-		}
-		return nullptr;
+		assert(it != std::end(_identifierToNodeItem));
+		return it->first;
 	}
 
 	void customContextMenuRequested(QPoint const& pos)
 	{
 		Q_Q(ControlledEntityTreeWidget);
 
-		auto const item = static_cast<NodeItem*>(q->itemAt(pos));
+		auto const* item = static_cast<NodeItem*>(q->itemAt(pos));
 		if (!item)
 		{
 			return;
 		}
 
-		auto const* node = find(item);
-		if (!node)
+		auto const nodeIdentifier = findNodeIdentifier(item);
+		if (nodeIdentifier.type == la::avdecc::entity::model::DescriptorType::Configuration)
 		{
-			return;
-		}
-
-		if (node->descriptorType == la::avdecc::entity::model::DescriptorType::Configuration)
-		{
-			auto const* configurationNode = static_cast<la::avdecc::controller::model::ConfigurationNode const*>(node);
+			auto const& anyNode = item->data(0, Qt::UserRole).value<AnyNode>().getNode();
+			auto const* configurationNode = std::any_cast<la::avdecc::controller::model::ConfigurationNode const*>(anyNode);
 
 			QMenu menu;
 
@@ -378,36 +399,25 @@ public:
 	}
 
 private:
-	la::avdecc::controller::model::Node const* find(NodeItem const* item)
+	NodeIdentifier makeIdentifier(la::avdecc::controller::model::Node const* node, Kind const kind) const noexcept
 	{
-		for (auto const& it : _map)
-		{
-			if (it.second == item)
-			{
-				return it.first;
-			}
-		}
-
-		return nullptr;
-	}
-
-	NodeItem* find(la::avdecc::controller::model::Node const* const node)
-	{
-		return node ? _map.at(node) : nullptr;
+		return { node->descriptorType, static_cast<la::avdecc::controller::model::EntityModelNode const*>(node)->descriptorIndex, kind };
 	}
 
 	template<typename T>
-	NodeItem* addItem(la::avdecc::controller::model::Node const* parent, T const* node, QString const& name) noexcept
+	NodeItem* addItem(la::avdecc::controller::model::Node const* parent, Kind const parentKind, T const* node, QString const& name) noexcept
 	{
 		NodeItem* item = nullptr;
 
 		if constexpr (std::is_base_of_v<la::avdecc::controller::model::EntityModelNode, T>)
 		{
 			item = new EntityModelNodeItem{ static_cast<la::avdecc::controller::model::EntityModelNode const*>(node), name };
+			_identifierToNodeItem.insert({ makeIdentifier(node, Kind::EntityModelNode), item });
 		}
 		else if constexpr (std::is_base_of_v<la::avdecc::controller::model::VirtualNode, T>)
 		{
 			item = new VirtualNodeItem{ static_cast<la::avdecc::controller::model::VirtualNode const*>(node), name };
+			_identifierToNodeItem.insert({ makeIdentifier(node, Kind::VirtualNode), item });
 		}
 		else
 		{
@@ -418,8 +428,10 @@ private:
 		auto const anyNode = AnyNode(node);
 		item->setData(0, Qt::UserRole, QVariant::fromValue(anyNode));
 
-		if (auto* parentItem = find(parent))
+		if (parent)
 		{
+			auto* parentItem = findItem(makeIdentifier(parent, parentKind));
+			assert(parentItem);
 			parentItem->addChild(item);
 		}
 		else
@@ -427,8 +439,6 @@ private:
 			Q_Q(ControlledEntityTreeWidget);
 			q->addTopLevelItem(item);
 		}
-
-		_map.insert({ node, item });
 
 		return item;
 	}
@@ -439,7 +449,7 @@ private:
 		{
 			return QString("%1: %2").arg(avdecc::helper::descriptorTypeToString(type), name);
 		};
-		auto* item = addItem(nullptr, &node, genName(node.dynamicModel->entityName.data()));
+		auto* item = addItem(nullptr, Kind::EntityModelNode, &node, genName(node.dynamicModel->entityName.data()));
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::entityNameChanged, item,
 			[genName, item](la::avdecc::UniqueIdentifier const entityID, QString const& entityName)
@@ -458,7 +468,7 @@ private:
 		{
 			return QString("%1.%2: %3").arg(avdecc::helper::descriptorTypeToString(type), QString::number(index), name);
 		};
-		auto* item = addItem(parent, &node, genName(avdecc::helper::configurationName(controlledEntity, node)));
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, genName(avdecc::helper::configurationName(controlledEntity, node)));
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::configurationNameChanged, item,
 			[this, genName, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, QString const& /*configurationName*/)
@@ -512,7 +522,7 @@ private:
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::AudioUnitNode const& node) noexcept override
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::audioUnitNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::AudioUnitIndex const audioUnitIndex, QString const& /*audioUnitName*/)
@@ -521,10 +531,10 @@ private:
 			});
 	}
 
-	void processStreamInputNode(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::Node const* const parent, la::avdecc::controller::model::StreamInputNode const& node) noexcept
+	void processStreamInputNode(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::Node const* const parent, Kind const parentKind, la::avdecc::controller::model::StreamInputNode const& node) noexcept
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, parentKind, &node, name);
 
 		auto& manager = avdecc::ControllerManager::getInstance();
 		auto const flags = manager.getStreamInputErrorCounterFlags(_controlledEntityID, node.descriptorIndex);
@@ -542,14 +552,14 @@ private:
 		// Only show non-redundant streams when the parent is Configuration
 		if (!node.isRedundant)
 		{
-			processStreamInputNode(controlledEntity, parent, node);
+			processStreamInputNode(controlledEntity, parent, Kind::EntityModelNode, node);
 		}
 	}
 
-	void processStreamOutputNode(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::Node const* const parent, la::avdecc::controller::model::StreamOutputNode const& node) noexcept
+	void processStreamOutputNode(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::Node const* const parent, Kind const parentKind, la::avdecc::controller::model::StreamOutputNode const& node) noexcept
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, parentKind, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::streamNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::DescriptorType const descriptorType, la::avdecc::entity::model::StreamIndex const streamIndex, QString const& /*streamName*/)
@@ -563,14 +573,14 @@ private:
 		// Only show non-redundant streams when the parent is Configuration
 		if (!node.isRedundant)
 		{
-			processStreamOutputNode(controlledEntity, parent, node);
+			processStreamOutputNode(controlledEntity, parent, Kind::EntityModelNode, node);
 		}
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::AvbInterfaceNode const& node) noexcept override
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::avbInterfaceNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::AvbInterfaceIndex const avbInterfaceIndex, QString const& /*avbInterfaceName*/)
@@ -582,25 +592,25 @@ private:
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::LocaleNode const& node) noexcept override
 	{
 		auto const name = QString("%1.%2: %3").arg(avdecc::helper::descriptorTypeToString(node.descriptorType), QString::number(node.descriptorIndex), node.staticModel->localeID.data());
-		addItem(parent, &node, name);
+		addItem(parent, Kind::EntityModelNode, &node, name);
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandParent*/, la::avdecc::controller::model::LocaleNode const* const parent, la::avdecc::controller::model::StringsNode const& node) noexcept override
 	{
 		auto const name = QString("%1.%2").arg(avdecc::helper::descriptorTypeToString(node.descriptorType), QString::number(node.descriptorIndex));
-		addItem(parent, &node, name);
+		addItem(parent, Kind::EntityModelNode, &node, name);
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandParent*/, la::avdecc::controller::model::AudioUnitNode const* const parent, la::avdecc::controller::model::StreamPortNode const& node) noexcept override
 	{
 		auto const name = QString("%1.%2").arg(avdecc::helper::descriptorTypeToString(node.descriptorType), QString::number(node.descriptorIndex));
-		addItem(parent, &node, name);
+		addItem(parent, Kind::EntityModelNode, &node, name);
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandGrandParent*/, la::avdecc::controller::model::AudioUnitNode const* const /*grandParent*/, la::avdecc::controller::model::StreamPortNode const* const parent, la::avdecc::controller::model::AudioClusterNode const& node) noexcept override
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::audioClusterNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::ClusterIndex const audioClusterIndex, QString const& /*audioClusterName*/)
@@ -612,13 +622,13 @@ private:
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandGrandParent*/, la::avdecc::controller::model::AudioUnitNode const* const /*grandParent*/, la::avdecc::controller::model::StreamPortNode const* const parent, la::avdecc::controller::model::AudioMapNode const& node) noexcept override
 	{
 		auto const name = QString("%1.%2").arg(avdecc::helper::descriptorTypeToString(node.descriptorType), QString::number(node.descriptorIndex));
-		addItem(parent, &node, name);
+		addItem(parent, Kind::EntityModelNode, &node, name);
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::ClockDomainNode const& node) noexcept override
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::clockDomainNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::ClockDomainIndex const clockDomainIndex, QString const& /*clockDomainName*/)
@@ -630,7 +640,7 @@ private:
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandParent*/, la::avdecc::controller::model::ClockDomainNode const* const parent, la::avdecc::controller::model::ClockSourceNode const& node) noexcept override
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::clockSourceNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::ClockSourceIndex const clockSourceIndex, QString const& /*clockSourceName*/)
@@ -642,7 +652,7 @@ private:
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::MemoryObjectNode const& node) noexcept override
 	{
 		auto const name = genName(controlledEntity, node);
-		auto* item = addItem(parent, &node, name);
+		auto* item = addItem(parent, Kind::EntityModelNode, &node, name);
 
 		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::memoryObjectNameChanged, item,
 			[this, item, node](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::MemoryObjectIndex const memoryObjectIndex, QString const& /*memoryObjectName*/)
@@ -654,17 +664,17 @@ private:
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::RedundantStreamNode const& node) noexcept override
 	{
 		auto const name = QString("REDUNDANT_%1.%2").arg(avdecc::helper::descriptorTypeToString(node.descriptorType), QString::number(node.virtualIndex));
-		addItem(parent, &node, name);
+		addItem(parent, Kind::EntityModelNode, &node, name);
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandParent*/, la::avdecc::controller::model::RedundantStreamNode const* const parent, la::avdecc::controller::model::StreamInputNode const& node) noexcept override
 	{
-		processStreamInputNode(controlledEntity, parent, node);
+		processStreamInputNode(controlledEntity, parent, Kind::VirtualNode, node);
 	}
 
 	virtual void visit(la::avdecc::controller::ControlledEntity const* const controlledEntity, la::avdecc::controller::model::ConfigurationNode const* const /*grandParent*/, la::avdecc::controller::model::RedundantStreamNode const* const parent, la::avdecc::controller::model::StreamOutputNode const& node) noexcept override
 	{
-		processStreamOutputNode(controlledEntity, parent, node);
+		processStreamOutputNode(controlledEntity, parent, Kind::VirtualNode, node);
 	}
 
 private:
@@ -672,10 +682,18 @@ private:
 	Q_DECLARE_PUBLIC(ControlledEntityTreeWidget);
 
 	la::avdecc::UniqueIdentifier _controlledEntityID{};
-	std::unordered_map<la::avdecc::controller::model::Node const*, NodeItem*> _map;
 
-	using NodeExpandedStates = std::unordered_map<la::avdecc::controller::model::Node const*, bool>;
-	std::unordered_map<la::avdecc::UniqueIdentifier, NodeExpandedStates, la::avdecc::UniqueIdentifier::hash> _entityExpandedStates;
+	// Quick access node item by node identifier
+	std::unordered_map<NodeIdentifier, NodeItem*, NodeIdentifier::hash> _identifierToNodeItem;
+
+	struct UserTreeWidgetState
+	{
+		NodeIdentifier currentNode{};
+		NodeIdentifierSet expandedNodes{};
+	};
+
+	// Global map that stores UserTreeWidgetState by entity ID (Stored in the class so each instance keeps its own cache)
+	std::unordered_map<la::avdecc::UniqueIdentifier, UserTreeWidgetState, la::avdecc::UniqueIdentifier::hash> _userTreeWidgetStates{};
 };
 
 ControlledEntityTreeWidget::ControlledEntityTreeWidget(QWidget* parent)